SPM locally added to Xcode project
Honestly don't know if this will hold up between branch changes and can't see how it would work between machines.

This is the result of dragging the KMP project set up to build the XCFramework and with the `Package.swift` and the base level to allow for all submodules to be explored in Xcode.